kok-foong/1/exposure-calculation-agent: moved a few functions to shared_utils.py
